Rename TemplateForm to JobTemplateForm
