+++ awx/ui_next/src/screens/Template/shared/index.js
@@ -1,1 +1,1 @@
-export { default } from './TemplateForm';
+export { default } from './JobTemplateForm';
